requirements.txt
absl-py==2.3.1
accelerate==1.10.1
adagio==0.2.6
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiohttp-cors==0.8.1
aiosignal==1.4.0
alembic==1.16.5
annotated-types==0.7.0
anyio==4.11.0
appdirs==1.4.4
async-timeout==5.0.1
attrs==25.3.0
autogluon.common==1.4.0
autogluon.core==1.4.0
autogluon.features==1.4.0
autogluon.tabular==1.4.0
autogluon.timeseries==1.4.0
bcrypt==5.0.0
boto3==1.40.41
botocore==1.40.41
cachetools==6.2.0
catboost==1.2.8
certifi==2025.8.3
charset-normalizer==3.4.3
chronos-forecasting==1.5.3
click==8.3.0
cloudpickle==3.1.1
colorful==0.5.7
colorlog==6.9.0
contourpy==1.3.2
coreforecast==0.0.16
cycler==0.12.1
distlib==0.4.0
dnspython==2.8.0
einops==0.8.1
email-validator==2.3.0
exceptiongroup==1.3.0
fastapi==0.118.0
fastapi-cli==0.0.13
fastapi-cloud-cli==0.2.1
filelock==3.13.1
fonttools==4.60.1
frozenlist==1.7.0
fs==2.4.16
fsspec==2024.6.1
fugue==0.9.1
future==1.0.0
gluonts==0.16.2
google-api-core==2.25.1
google-auth==2.41.0
googleapis-common-protos==1.70.0
graphviz==0.21
greenlet==3.2.4
grpcio==1.75.1
h11==0.16.0
hf-xet==1.1.10
httpcore==1.0.9
httptools==0.6.4
httpx==0.28.1
huggingface-hub==0.35.3
hyperopt==0.2.7
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
jmespath==1.0.1
joblib==1.5.2
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
kiwisolver==1.4.9
lightgbm==4.6.0
lightning==2.5.5
lightning-utilities==0.15.2
llvmlite==0.45.0
Mako==1.3.10
Markdown==3.9
markdown-it-py==4.0.0
MarkupSafe==2.1.5
matplotlib==3.10.6
mdurl==0.1.2
mlforecast==0.14.0
mpmath==1.3.0
msgpack==1.1.1
multidict==6.6.4
narwhals==2.6.0
networkx==3.3
numba==0.62.1
numpy==2.1.2
nvidia-cublas-cu12==12.8.3.14
nvidia-cuda-cupti-cu12==12.8.57
nvidia-cuda-nvrtc-cu12==12.8.61
nvidia-cuda-runtime-cu12==12.8.57
nvidia-cudnn-cu12==9.7.1.26
nvidia-cufft-cu12==11.3.3.41
nvidia-cufile-cu12==1.13.0.11
nvidia-curand-cu12==10.3.9.55
nvidia-cusolver-cu12==11.7.2.55
nvidia-cusparse-cu12==12.5.7.53
nvidia-cusparselt-cu12==0.6.3
nvidia-nccl-cu12==2.26.2
nvidia-nvjitlink-cu12==12.8.61
nvidia-nvtx-cu12==12.8.55
opencensus==0.11.4
opencensus-context==0.1.3
optuna==4.5.0
orjson==3.11.3
packaging==25.0
pandas==2.3.3
passlib==1.7.4
patsy==1.0.1
pillow==11.0.0
platformdirs==4.4.0
plotly==6.3.0
prometheus_client==0.23.1
propcache==0.3.2
proto-plus==1.26.1
protobuf==6.32.1
psutil==7.0.0
py-spy==0.4.1
py4j==0.10.9.9
pyarrow==20.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pydantic==2.11.9
pydantic-extra-types==2.10.5
pydantic-settings==2.11.0
pydantic_core==2.33.2
Pygments==2.19.2
PyJWT==2.10.1
pyparsing==3.2.5
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-multipart==0.0.20
pytorch-lightning==2.5.5
pytz==2025.2
PyYAML==6.0.3
ray==2.44.1
referencing==0.36.2
regex==2025.9.18
requests==2.32.5
rich==14.1.0
rich-toolkit==0.15.1
rignore==0.6.4
rpds-py==0.27.1
rsa==4.9.1
s3transfer==0.14.0
safetensors==0.6.2
scikit-learn==1.7.2
scipy==1.15.3
sentencepiece==0.2.1
sentry-sdk==2.39.0
shellingham==1.5.4
six==1.17.0
smart_open==7.3.1
sniffio==1.3.1
SQLAlchemy==2.0.43
starlette==0.48.0
statsforecast==2.0.1
statsmodels==0.14.5
sympy==1.14.0
tensorboard==2.20.0
tensorboard-data-server==0.7.2
tensorboardX==2.6.4
threadpoolctl==3.6.0
tokenizers==0.21.4
tomli==2.2.1
toolz==0.12.1
torch==2.7.1+cu128
torchaudio==2.7.1+cu128
torchmetrics==1.8.2
torchvision==0.22.1+cu128
tqdm==4.67.1
transformers==4.49.0
triad==0.9.8
triton==3.3.1
typer==0.19.2
typing-inspection==0.4.1
typing_extensions==4.15.0
tzdata==2025.2
ujson==5.11.0
urllib3==2.5.0
utilsforecast==0.2.11
uvicorn==0.37.0
uvloop==0.21.0
virtualenv==20.34.0
watchfiles==1.1.0
websockets==15.0.1
Werkzeug==3.1.3
window_ops==0.0.15
wrapt==1.17.3
xgboost==3.0.5
yarl==1.20.1
